requirements.txt
Flask==2.2.3
Flask-JWT-Extended==4.3.0
SQLAlchemy==1.4.25
mysql-connector-python==8.0.27
flask-sqlalchemy-3.0.3
flask-sqlalchemy==3.0.3
PyJWT == 2.6.0
